Find square root of a number

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -12,7 +12,7 @@
 int _sqrt_recursion(int n)
 {
 	const int num = n;
-	int sqrt = n;
+	long int sqrt = n;
 
 	if (sqrt * sqrt == num)
 	{
